move related structs togetherin controller
